+++ backend/src/main/java/com/softlanches/products/dto/ProductResponse.java
@@ -1,10 +1,12 @@
 package com.softlanches.products.dto;
 
+import java.math.BigDecimal;
+
 public record ProductResponse(
         Long id,
         String nome,
-        double preco,
-        double custo,
+        BigDecimal preco,
+        BigDecimal custo,
         boolean ativo,
         Long receitaId,
         double margem
