+++ backend/src/main/java/com/softlanches/recipes/dto/MapIngredientRequest.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
+
 public record MapIngredientRequest(
 
         @NotNull(message = "insumoId é obrigatório")
@@ -11,5 +13,5 @@ public record MapIngredientRequest(
         Long insumoId,
 
         @PositiveOrZero(message = "Fator de conversão não pode ser negativo")
-        Double fatorConversao
+        BigDecimal fatorConversao
 ) {}
