+++ backend/src/main/java/com/softlanches/recipes/dto/MappingResponse.java
@@ -1,8 +1,10 @@
 package com.softlanches.recipes.dto;
 
+import java.math.BigDecimal;
+
 public record MappingResponse(
         Long id,
         Long ingredienteId,
         Long insumoId,
-        Double fatorConversao
+        BigDecimal fatorConversao
 ) {}
